moved sampler
